refactor(mcp): extract BOM table-merge helpers into shared bom_table module (closes #850)
#833 left ``foundation/bom.py`` importing private helpers from ``prefab_ui``
to precompute ``extras["applied_plan_rows"]`` on the apply path — Copilot
flagged this as UI/impl coupling. The "BOM actions → merged row list /
outcome summary" computation is domain logic, not rendering.

Move the table-merge helpers into ``katana_mcp/tools/foundation/bom_table.py``
(non-UI, no Prefab imports). The renderer (``prefab_ui.build_bom_modify_ui``)
and the tool-impl (``foundation/bom._modify_product_bom_impl``) both import
from the new module — no more underscore-helper crossing the UI/impl
boundary.

Moved:

- ``_merge_bom_rows_for_modify_card`` + row-shaping helpers
  (``_bom_existing_row_from_snapshot``, ``_bom_synth_orphan_row``,
  ``_bom_add_action_to_row``, ``_apply_bom_update_to_row``,
  ``_apply_bom_delete_to_row``, ``_bom_change_lookup``)
- ``_prepare_bom_table_rows``
- ``_format_bom_quantity`` / ``_format_bom_quantity_diff``
- ``_BOM_ROW_STATUS_VARIANTS`` + ``_BomMergedRowKind``
- ``_summarize_apply_outcome`` (cross-entity-generic, but only directly
  called by ``foundation/bom`` and ``prefab_ui.build_po_modify_ui``;
  housing it here lets ``bom.py`` import it without a UI dep — small
  pragmatic call, hoist to a dedicated shared-modify module later if a
  third caller appears)
- ``_derive_status_label`` (dict version; the ``ActionResult`` version
  in ``_modification.py`` is untouched)

Kept in ``prefab_ui.py``:

- ``build_bom_modify_ui`` and all rendering glue (public render contract)
- ``_bom_row_summary`` (compose the "+N added, ~M updated, -K deleted"
  display line — purely presentational)
- ``_BOM_MODIFY_COLUMNS`` + ``_resolve_bom_table_rows`` (UI-bound
  configuration / wiring)

Test impact: ``test_unknown_operation_is_logged_and_dropped`` updates the
caplog logger name from ``katana_mcp.tools.prefab_ui`` to
``katana_mcp.tools.foundation.bom_table`` (the logger followed the code).
All other tests pass unchanged — they import the moved symbols from
``prefab_ui`` which re-exports them transparently via its own import.
